error handling

+++ Ewaste.py
@@ -6,16 +6,19 @@ import time
 
 FolderPath = input("Full Path to the Ewaste folder, i.e. (C:/Users/Username/path/to/Ewaste/): ")
 
-Contents = os.listdir(FolderPath)
-
-DelConfirm = input(f"You are about to remove these files: {Contents} Are you sure [Y/N]: ")
-
-if DelConfirm == "y":
-    shutil.rmtree(FolderPath)
-    print("Deleted your Ewaste. Exiting soon please do not close.")
-    time.sleep(3)
-    sys.exit()
+if "Ewaste" in FolderPath:
+    Contents = os.listdir(FolderPath)
+    
+    DelConfirm = input(f"You are about to remove these files: {Contents} Are you sure [Y/N]: ")
+    
+    if DelConfirm == "y":
+        shutil.rmtree(FolderPath)
+        print("Deleted your Ewaste. Exiting soon please do not close.")
+        time.sleep(3)
+        sys.exit()
+    else:
+        print("Okay. Exiting soon please do not close.")
+        time.sleep(3)
+        sys.exit()
 else:
-    print("Okay. Exiting soon please do not close.")
-    time.sleep(3)
-    sys.exit()
+    print("An error occured. Please make an Ewaste folder with your EWaste in it, and feed this program the path.")
